Sbt process testrunner (#492)
### Closes #152 

Changes the sbt-plugin to run tests in a separate process. The plugin will start a new process with the new sbt-testrunner project and the application to test on the classpath. Also adds an api project to describe messages that are sent between the main and sub-processes.

The sbt-testrunner project will programmatically start tests using the [sbt test-interface API](https://github.com/sbt/test-interface). I made a small [POC to run tests programmatically](https://github.com/hugo-vrijswijk/test-init-poc) with the interface using either ScalaTest or MUnit. Both of those frameworks (and any testing frameworks that work in sbt) implement the test-interface.

Out of scope for this PR:

- Plugins that aren't sbt. 
  - This is probably possible for the maven plugin, but I want a working MVP first. And it would likely require additional setup for junit tests or other test frameworks as those don't implement the sbt test-interface
- Setups with multiple testing frameworks. Assumption is a single test-framework. Will probably work, but will be verified after this PR.
- Detecting static mutants (in objects, for example, the code will be run before any mutations are activated, so they can never be mutated). We should ignore these in the future
- Filtering tests on coverage
- Filtering tests on speed
- Starting multiple processes for performance
- Scala.js. JVM only for now

It also opens up the way towards features like #412, #312, #16 and #10 (wow that's a big list)