Refactor file reading
File reading is now completely done with fs2-io, and the dependency on better-files is removed. Additionally, a sbt file resolver was added that resolves files using sbt settings, instead of globs. In many cases, this will be the preferred way of reading files. When setting `mutate` or `files`, those are still used.

Because of fs2, file globbing, reading and parsing is now also done in parallel, using the default concurrency from the config