Refactor file reading (#929)
File reading is majorly refactored to now completely use fs2-io, and the dependency on better-files is removed. The FileCollector is gone and split into MutatesFileResolver and FilesFileResolver corresponding to mutate and files options, respectively.

Additionally, a sbt file resolver was added that resolves files using sbt settings, instead of globs. In many cases, this will be the preferred way of reading files. When setting mutate or files, those are still used.

Because of fs2, file globbing, reading and parsing is now also done in parallel, using the default concurrency from the config. This should make Stryker4s faster before starting with running tests.